moved unused files, renamed functions, change attention layer adding
