+++ src/layers/attention.py
@@ -1,0 +1,24 @@
+from keras.layers import Layer
+from keras import backend as K
+
+class attention(Layer):
+    def __init__(self, return_sequences=True):
+        self.return_sequences = return_sequences
+        super(attention,self).__init__()
+
+    def build(self, input_shape):
+        self.W=self.add_weight(name="att_weight", shape=(input_shape[-1],1),
+                               initializer="normal")
+        self.b=self.add_weight(name="att_bias", shape=(input_shape[1],1),
+                               initializer="normal")
+        super(attention,self).build(input_shape)
+
+
+    def call(self, x):
+        print("called")
+        e = K.tanh(K.dot(x,self.W)+self.b)
+        a = K.softmax(e, axis=1)
+        output = x*a
+        if self.return_sequences:
+            return output
+        return K.sum(output, axis=1)
