+++ src/main.py
@@ -3,20 +3,21 @@ import time as tm
 import datetime as dt
 from yahoo_fin import stock_info as yf
 import numpy as np
-from models.Lstm.lstm import lstm_main
-from models.LstmAttention.lstmAttention import lstmAttentionMain
-def getTimeInterval():
+from models.lstm.lstm import lstm_main
+from models.bidirectionallstm.bilstm import bi_lstm_main
+
+def get_time_interval():
     date_now = tm.strftime('%Y-%m-%d')
     date_3_years_back = ((dt.date.today() - dt.timedelta(days=2)) - dt.timedelta(days=1200)).strftime('%Y-%m-%d')
     return date_now, date_3_years_back
 
-def scaleData(data):
+def scale_data(data):
     scaler = MinMaxScaler()
     data['close'] = scaler.fit_transform(np.expand_dims(data['close'].values, axis=1))
     return data
 
-def getCleanData(stock):
-    date_now,date_3_years_back = getTimeInterval()
+def get_clean_data(stock):
+    date_now,date_3_years_back = get_time_interval()
     init_df = yf.get_data(
     stock, 
     start_date=date_3_years_back, 
@@ -31,23 +32,23 @@ def main():
     stock = 'AMZN'
 
     scaler = MinMaxScaler()
-    data = getCleanData(stock)
+    data = get_clean_data(stock)
     data_visualisation = data.copy()
     data['close'] = scaler.fit_transform(np.expand_dims(data['close'].values, axis=1))
     raw_seq = data['close']
 
     #input is data - pred_days - runs
-    scaled_lstm =  lstm_main(raw_seq,2,1)
-    descaled_lstm = []
-    for avg in scaled_lstm:
-        descaled_lstm.append(scaler.inverse_transform(np.array(avg).reshape(-1,1))[0][0])
+    # scaled_lstm =  lstm_main(raw_seq,1,1,True)
+    # descaled_lstm = []
+    # for avg in scaled_lstm:
+    #     descaled_lstm.append(scaler.inverse_transform(np.array(avg).reshape(-1,1))[0][0])
 
-    scaled_lstm_attention =  lstmAttentionMain(raw_seq,2,1)
-    descaled_lstm_attention = []
-    for avg in scaled_lstm_attention:
-        descaled_lstm_attention.append(scaler.inverse_transform(np.array(avg).reshape(-1,1))[0][0])
+    scaled_bi_lstm =  bi_lstm_main(raw_seq,2,1,False)
+    descaled_bi_lstm= []
+    for avg in scaled_bi_lstm:
+        descaled_bi_lstm.append(scaler.inverse_transform(np.array(avg).reshape(-1,1))[0][0])
 
-    print(descaled_lstm)
-    print(descaled_lstm_attention)
+    #print(descaled_lstm)
+    print(descaled_bi_lstm)
 main()
